+++ backend/products.py
@@ -2,9 +2,17 @@ from __main__ import app, db
 from flask import request, jsonify
 
 
-@app.route("/get_products", methods=["GET"])
+@app.route("/get_products", methods=["POST"])
 def get_products():
+    encrypted_session_str = request.json["session_string"]
+    session_str = decrypt_session_string(encrypted_session_str).split(";")
+
     cursor = db.cursor()
+
+    cursor.execute("select * from Users where email = %s and password = %s and role = %s", (session_str[0], session_str[1], session_str[2]))
+    is_admin = cursor.fetchone()
+    if not is_admin:
+        return jsonify({"status": "fail", "message": "Invalid session string, not an admin account"})
 
     cursor.execute("select * from Products")
     products = cursor.fetchall()
@@ -24,11 +32,18 @@ def get_products():
     return jsonify({"status": "success", "products": product_list})
 
 
-@app.route("/get_product", methods=["GET"])
+@app.route("/get_product", methods=["POST"])
 def get_product():
-    product_id = request.args.get("id")
+    product_id = request.json["id"]
+    encrypted_session_str = request.json["session_string"]
+    session_str = decrypt_session_string(encrypted_session_str).split(";")
 
     cursor = db.cursor()
+
+    cursor.execute("select * from Users where email = %s and password = %s and role = %s", (session_str[0], session_str[1], session_str[2]))
+    is_admin = cursor.fetchone()
+    if not is_admin:
+        return jsonify({"status": "fail", "message": "Invalid session string, not an admin account"})
 
     cursor.execute("select * from Products where id = %s", (product_id,))
     product = cursor.fetchone()
@@ -104,11 +119,18 @@ def edit_product():
         return jsonify({"status": "success", "message": "Product updated successfully!"})
 
 
-@app.route("/delete_product", methods=["DELETE"])
+@app.route("/delete_product", methods=["POST"])
 def delete_product():
-    product_id = request.args.get("id")
+    product_id = request.json["id"]
+    encrypted_session_str = request.json["session_string"]
+    session_str = decrypt_session_string(encrypted_session_str).split(";")
 
     cursor = db.cursor()
+
+    cursor.execute("select * from Users where email = %s and password = %s and role = %s", (session_str[0], session_str[1], session_str[2]))
+    is_admin = cursor.fetchone()
+    if not is_admin:
+        return jsonify({"status": "fail", "message": "Invalid session string, not an admin account"})
 
     cursor.execute("delete from Products where id = %s", (product_id,))
     db.commit()
